% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryData.R
\name{numeric_summary}
\alias{numeric_summary}
\title{Generate summary statistics}
\usage{
numeric_summary(data, column_name)
}
\arguments{
\item{data}{A data frame containing the column to be summarized (e.g., gapminder dataset). I named this data because it seems to be a
very common name for the object even for other functions in other packages.}

\item{column_name}{The name of the column in the data frame to be summarized.}
}
\value{
The count, mean, median, and standard deviation of numeric values.
}
\description{
This function calculates the different summary statistics (eg.count, mean, median, and standard deviation) of numeric values
in the specified column of a data frame. In the case of non-numeric values, it would at least get counts.
}
\examples{
library(gapminder)
numeric_summary(gapminder, "pop")
numeric_summary(gapminder, "gdpPercap")

}
